[graphite] Upload dither range uniform as a Half
Bug: b/256870077
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/661040
Commit-Queue: Robert Phillips <robertphillips@google.com>

+++ src/gpu/graphite/ShaderCodeDictionary.cpp
@@ -737,7 +737,7 @@ void GenerateCoordClampPreamble(const ShaderInfo& shaderInfo,
 
 //--------------------------------------------------------------------------------------------------
 static constexpr Uniform kDitherShaderUniforms[] = {
-        { "range", SkSLType::kFloat },
+        { "range", SkSLType::kHalf },
 };
 
 static constexpr TextureAndSampler kDitherTexturesAndSamplers[] = {
